Don't attempt to dump memory segments marked VM_DONTDUMP
In newer versions of the Linux kernel, some parts of the vvar memory
segment are not readable by userspace.  In this situation the core
dump write will fail when we get an EFAULT error attempting to read
this memory.

This memory is not needed by the debugger, so we omit it by looking at
the VmFlags for each memory segment to see if the "dd" (VM_DONTDUMP)
flag is set; if so we don't write that segment to the core file.