Resolve unittest crash with newer GCC at -O1
If I compiled with GCC 8.1 and -O1 (or -Og) I was getting core dumps in
the coredump_unittest.  After much experimentation I discovered that
the builtin memcpy() in one specific place in elfcore.c was causing
this.  Replacing that with a trivial local memcpy() (we can still use
builtin memcpy elsewhere) fixes the issue.

I don't know why: there's nothing magical about this memcpy as far as
I can see.