Introduced an API to collect the core from an unrelated process
The API WritePidCoreDumpWith() will take argument pid, and try to
collect the core from the process with pid, this uses the
existing logic to identify the threads under the given process
and to identify the maps and the properties of each memory maps.
Then read the memory content from /proc/[pid]/mem, we assume that
the caller has the permission to trace and read the mem from
process [pid].

A tool also introduced, which will use the API WritePidCoreDumpWith
and provide a convenient CLI for the user, which will help in
creating core dump of any process that the user has permission to trace.