Introduced an API to collect the core from an unrelated process
The API WriteCoreDumpOfPID() will take argument pid, and try to
collect the core from the process with pid, this uses the
existing logic to identify the threads under the given process
and to identify the maps and the properties of each memory maps.
Then read the memory content from /proc/[pid]/mem, we assume that
the caller has the permission to trace and read the mem from
process [pid].

A tool also introduced, which will use the API WriteCoreDumpOfPID
and provide a convenient CLI for the user, which will help in
creating core dump of any process that the user has permission to trace.

Following is the help from the tool:
Usage: ./coredumper_tool <-c corefile> <-p pid> [-s core_size] [-z]
<-h>               Show this message and exit
<-c core_file>     Collect the core dump into a file core_file
<-p pid>           Collect the core dump of process with pid
<-s core_size>     Limit the size of core_file to core_size
<-z>               Compress the core_file on the fly with gzip